Geometry file validation in its own class
