+++ nexus_constructor/geometry_file_validator.py
@@ -1,17 +1,16 @@
 from PySide2.QtCore import QObject, QUrl, Slot
-import sys
+
 
 class GeometryFileValidator(QObject):
     @Slot(QUrl, result=bool)
     def validate_geometry_file(self, file_url: QUrl):
 
-        ext = file_url.toString(options=QUrl.FormattingOptions(QUrl.PreferLocalFile))[
-            -3:
-        ].lower()
-        sys.stdout.write(ext)
-        if ext == "off":
-            return False
-        elif ext == "stl":
-            return False
+        ext = file_url.toString(
+            options=QUrl.FormattingOptions(QUrl.PreferLocalFile)
+        ).lower()[-3:]
 
-        return True
+        if ext == "off":
+            pass
+
+        elif ext == "stl":
+            pass
